requirements.txt
maya
PyYaml
markdown